internal/fs: use build tags to define the behaviour of fs.RenameWithFallback
